Update CircularLinkedList.java
A linked list in which the last node points to the first node is called circular linked list. The nodes are circularly connected. There is no NULL at the end of list. 

Types: Single and Double Linked Lists

Benefits: 
1. Nodes can be tranversed circularly i.e. we can reach to any node from any node. 
2. Only one pointer is required unlike circular queue where two pointers are required. A pointer is kept at the last node of the list. The node next to the last node is the first node of the list. 

Application:

Contact list - When the contact list is scrolled, the contacts starting with A reappear after contacts starting with Z. It is an implementation of double circular linked list.